requirements.txt
azure-functions
pydocumentdb